96. Pruebas - SaveFile UseCase custom values

+++ src/domain/use-cases/save-file.use-case.test.ts
@@ -2,8 +2,25 @@ import fs from 'fs';
 import { SaveFile } from './save-file.use-case';
 
 describe('SaveFileUseCase', () => {
+	const customOptions = {
+		fileContent: 'custom content',
+		fileDestination: 'custom-outputs/file-destination',
+		fileName: 'custom-table-name'
+	};
+
+	const customFilePath = `${customOptions.fileDestination}/${customOptions.fileName}.txt`;
+
 	afterEach(() => {
-		fs.rmSync('outputs', { recursive: true });
+		const outputFolderExists = fs.existsSync('outputs');
+
+		if (outputFolderExists) fs.rmSync('outputs', { recursive: true });
+
+		const customOutputFolderExists = fs.existsSync(
+			customOptions.fileDestination
+		);
+
+		if (customOutputFolderExists)
+			fs.rmSync(customOptions.fileDestination, { recursive: true });
 	});
 
 	test('should save file with default values', () => {
@@ -26,5 +43,21 @@ describe('SaveFileUseCase', () => {
 		expect(fileExists).toBe(true);
 
 		expect(fileContent).toBe(options.fileContent);
+	});
+
+	test('should save file with custom values', () => {
+		const saveFile = new SaveFile();
+
+		const result = saveFile.execute(customOptions);
+
+		const fileExists = fs.existsSync(customFilePath);
+
+		const fileContent = fs.readFileSync(customFilePath, { encoding: 'utf-8' });
+
+		expect(result).toBe(true);
+
+		expect(fileExists).toBe(true);
+
+		expect(fileContent).toBe(customOptions.fileContent);
 	});
 });
